<Enhancement>[Preferences]: Remove Bends
While doing this change fix the docs, introduce type hinting where it is missing;  Simple unit test for new functionality

[https://github.com/hasii2011/PyUt/issues/203]

+++ src/org/pyut/miniogl/PointShape.py
@@ -14,7 +14,7 @@ class PointShape(Shape):
     A point, which is drawn as a little square (3 pixels wide).
 
     """
-    def __init__(self, x: float, y: float, parent=None):
+    def __init__(self, x: int, y: int, parent=None):
         """
 
         Args:
